backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
sqlmodel==0.0.6
python-dotenv==0.20.0
python-multipart==0.0.5
requests==2.27.1
Faker==13.12.1
arrow==1.2.2
sqlalchemy-utils==0.38.2
pg8000==1.29.1
